feat(Movie): Movie 도메인엔티티 구현
